requirements.txt
playwright>=1.40.0
pandas>=2.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
asyncio>=3.4.3
pydantic>=2.0.0
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-mock>=3.11.0
pyyaml>=6.0
